% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{cr_interpolate}
\alias{cr_interpolate}
\title{Interpolate the Disease Mortality Data}
\usage{
cr_interpolate(data, reliable = TRUE)
}
\arguments{
\item{data}{original data. It is assumed to have a \code{Deaths}, \code{Population}, and \code{Crude.Rate} column.}

\item{reliable}{if taking into account deaths <= 20. (Default is \code{TRUE})}
}
\value{
a data frame with interpolated mortality rate
}
\description{
In the original data from CDC WONDER Online Database, the mortality rate in counties with less than 20 cases
of deaths were recorded as "unreliable". In this function, users may choose whether to interpolate the unreliable
values. If reliability is TRUE, the unreliable information is kept, and the string of "unreliable" will be
substituted with NA. Otherwise, the value will be interpolated by calculating the quotient of number of deaths
divided by the population.
}
\examples{
data("diabetes")
diabetes |> cr_interpolate(reliable = TRUE)
}
