% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes Deaths per County from CDC}
\description{
\code{County} is name of county and abbreviation of state name; \cr
\code{County.Code} is the FIPS code for the county; \cr
\code{Deaths} is number of people died because of diabetes in this county; \cr
\code{Population} is the population of the county; \cr
\code{Crude.Rate} is number of people died because of diabetes per 100,000 people in the county;
}
\keyword{data}
