% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{mnnn_to_na}
\alias{mnnn_to_na}
\title{Change NA to Target Value}
\usage{
mnnn_to_na(data, column, target)
}
\arguments{
\item{data}{original data.}

\item{column}{column needs to be substituted.}

\item{target}{target value.}
}
\value{
a data frame with NA's of selected columns replaced with the target value.
}
\description{
This function changes NA in the selected data columns to the target value.
}
\examples{
data("diabetes")
diabetes |> mnnn_to_na("Crude.Rate", "Unreliable")
}
