% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mortality_map}
\alias{mortality_map}
\title{Get the Spectral Signature of Accelerometry Data}
\usage{
mortality_map(data, colname, level = "County", disease = "Diabetes")
}
\arguments{
\item{data}{data}

\item{colname}{string of column name which represents the mortality in the data}

\item{level}{string, "County" or "State"}

\item{disease}{string of the disease name showing}
}
\value{
a plot
}
\description{
The spectral signature is calculated by taking the modulus of the
Fourier coefficients of the signal.
}
