% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mortality_vs_svi_scatter}
\alias{mortality_vs_svi_scatter}
\title{Get the Spectral Signature of Accelerometry Data}
\usage{
mortality_vs_svi_scatter(data, index, disease = "Diabetes")
}
\arguments{
\item{data}{prepared data}

\item{disease}{string of disease name}

\item{colname}{x}
}
\value{
a plot
}
\description{
The spectral signature is calculated by taking the modulus of the
Fourier coefficients of the signal.
}
