% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{prepare}
\alias{prepare}
\title{Get the Spectral Signature of Accelerometry Data}
\usage{
prepare(vulnerability, mortality, reliable = TRUE)
}
\arguments{
\item{vulnerability}{original SVI data from CDC}

\item{mortality}{original mortality data from CDC}

\item{reliable}{if taking into account case<= 20 (Default is \code{TRUE})}
}
\value{
a data frame
}
\description{
The spectral signature is calculated by taking the modulus of the
Fourier coefficients of the signal.
}
