\name{SVI-package}
\alias{SVI-package}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
SVI: Stochastic Variational Inference
}
\description{
This package is an implementation of the Stochastic Variational Inference to compute an approximate posterior. By taking an advantage of the mean field family and natural gradient, it achieves computational efficiency. Also it can handle a large scale data set.
}


\references{
1.Matthew D. Hoffman, David M. Blei, Chong Wang, John Paisley. Stochastic Variational Inference. \emph{Journal of Machine Learning Research}, 2013\cr
2.Nicholas J. Foti, Jason Xu, Dillon Laird, Emily B. Fox. Stochastic Variational Inference for Hidden Markov Models. \emph{Annual Conference on Neural Information Processing System}, 2014 \cr
}
\author{
Zhehui Chen, Shiyang Li, Xingguo Li, Tuo Zhao
}