% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVI_HMM.R
\name{SVI_HMM}
\alias{SVI_HMM}
\title{Solve the hyperparameters in HMM}
\usage{
SVI_HMM(Data_sequence, sub_lengthh, KK, W_AA, uu, kk, sigma, vv, U_AA, U_phi11, U_phi22,
  U_phi3, U_phi44,Dimm, Passs, pree)
}
\arguments{
\item{Data_sequence}{All the observed value, should be a list}

\item{sub_lengthh}{The length of sub-sequence used in each iteration}

\item{KK}{The number of hidden states}

\item{W_AA}{The initial of the variational parameter for transition matrix}

\item{uu}{The initial of the variational parameter for mean of NIW distribution}

\item{kk}{The initial of the variational parameter for k of NIW distribution}

\item{sigma}{The initial of the variational parameter for Vaiarance of NIW distribution}

\item{vv}{The initial of the variational parameter for v of NIW distribution.}

\item{U_AA}{The parameter of the prior for transition matrix}

\item{U_phi11}{The parameter of the prior for mean of NIW distribution}

\item{U_phi22}{The parameter of the prior for k of NIW distribution}

\item{U_phi3}{The parameter of the prior for Vaiarance of NIW distribution}

\item{U_phi44}{The parameter of the prior for v of NIW distribution}

\item{Dimm}{The Dimension of our data}

\item{Passs}{The number of passes we run, also the maximum number of iteration}

\item{pree}{Stop Criteria}
}
\value{
All variational parameters in the model.\item{Tran}{The hyperparameter of the transition matrix among hidden states.}\item{mu}{The mean value for each hidden state.}\item{kappa}{The parameter in NIW distribution for each hidden state}\item{Sigma}{The covariance matrix in NIW distribution for each hidden state.}\item{nu}{The parameter in the NIW distribution for each hidden state.}
}
\description{
In this function, we solve the hidden markov model through stochasitc variational inference. We use the sub-chains to cauculate the variaitional parameters and update them by the formula we get in the reference.
}
\examples{
## Generate data in the following way.
data <-list(c(1.8,2.0))
for(i in 1:10000)
{
 if(i\%\%6==0||i\%\%6==1)
 {
   data[[i]]=rnorm(2);
 }
 else
 {
   data[[i]]= rnorm(2,c(100,100),10);
 }
 ## initial our hyperparameters and give our prior parameters.
 K=2
 Dim=length(data[[1]])   ## Gassiuan Component Number
 sigma0 = diag(rep(1,2))
 u0 = 1:Dim # initial mean value
 k0=1             ## initial parameter
 v0=6             ## initial parameter
 ## initial covariance with identity matrix
 U_A = matrix(rep(1/K,K*K),K)        ## Hyperperameter the same notation with the paper
 U_phi1 = matrix(rep(0,Dim*K),Dim)  ## Hyperperameter the same notation with the paper
 for(i in 1:K)
 {
   U_phi1[,i]=2*k0*u0;
 }
 U_phi2 = rep(2*k0,K)            ##  Hyperperameter the same notation with the paper
 U_phi3 = list(K)

 for(i in 1:K)
 {
   U_phi3[[i]]=  sigma0+2*k0*u0\%*\%t(u0);       ## initial every component with the same covariance
 }
 U_phi4 = rep(v0+2+2,K)          ## Hyperperameter the same notation with the paper
 sub_length = 50                 ## sample chain length  # data_dimention
 u = matrix(rep(0,Dim*K),Dim) ;  ## claim mean value of K components.
 k = 1:K;                        ## claim k, which controls Gussuian Distribution
 v = (3+Dim):(2+Dim+K);          ## claim v, which controls Gussuian Distribution
 for( i in 1:K)
 {
   u[,i]=u0;                     ## initial every component with the same mean value
 }
 sigma = list(K)
 for(i in 1:K)
 {
   sigma[[i]]= sigma0;           ## initial every component with the same covariance
 }
 W_A = matrix(1:(K*K),K);
 }
## run the code
result = SVI_HMM(data,sub_length =9999,K=2,W_A, u, k, sigma, v, U_A, U_phi1,U_phi2, U_phi3, U_phi4,Dim, Pass=10000, pre=0.01)
}
\references{
Nicholas J. Foti, Jason Xu, Dillon Laird, Emily B. Fox. Stochastic Variational Inference for Hidden Markov Models. \emph{Annual Conference on Neural Information Processing System}, 2014 \cr
}
\author{
Zhehui Chen, Shiyang Li, Xingguo Li, Tuo Zhao
}
