% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVI_LDA.R
\name{SVI_LDA}
\alias{SVI_LDA}
\title{Solve the hyperparameters in LDA}
\usage{
SVI_LDA(X, KK, nn, alphaa, etaa, pree, word_lengthh)
}
\arguments{
\item{X}{The Input data, should be a list}

\item{KK}{The number of the topics}

\item{nn}{The number of passes we run, also the maximum number of iteration}

\item{alphaa}{The prior of theta, distribution of vocabulary over different topics}

\item{etaa}{The prior of beta, distribution of topics in each position over different document.}

\item{pree}{The stop criteria}

\item{word_lengthh}{The number of top words}
}
\value{
All the hyperparameters in the model. \item{Lambda}{The hyperparameter for each words of the dictionary in each topic.}\item{phi}{The hyperparameter for different topics of each position in each document.}\item{gamma}{The hyperparameter for topics proportion in each document.}\item{Topic}{The top words in each topic.}\item{Dictionary}{The dictionary for all documents except the stop words.}
}
\description{
In this function,we require users to input their data by a list, that is, each document is one vector as a component in the list. First, we remove the stop words in each document to make those top words more meaningful. Then we build a dictionary of these document. Finally, we implement the LDA with SVI. Since we initialize the latent parameters randomly, you may not get the same result even you have the same input.
}
\examples{
#####################################################################################
## test data
doc_a = c("Brocolli", "is", "good", "to", "eat", "My", "brother", "likes", "to", "eat", "good", "brocolli", "but", "not", "my", "mother")
doc_b = c("My", "mother", "spends", "a", "lot", "of", "time", "driving", "my", "brother", "around", "to", "baseball", "practice")
doc_c = c("Some", "health", "experts", "suggest", "that", "driving", "may cause", "increased", "tension", "and", "blood", "pressure")
doc_d = c("I", "often", "feel", "pressure", "to", "perform", "well", "at", "school", "but", "my", "mother", "never", "seems", "to", "drive", "my", "brother", "to", "do", "better")
doc_e = c("Health", "professionals", "say", "that", "brocolli", "is", "good", "for", "your", "health")
data <- list(doc_a,doc_b,doc_c,doc_d,doc_e)
## Initial prior parameters
K=3;
alpha=1/K;
eta=0.01;
n=1;
pre=0.01;
word_length=3;
## run the function
t=SVI_LDA(data,K,n,alpha,eta,pre,word_length)
}
\references{
Matthew D. Hoffman, David M. Blei, Chong Wang, John Paisley. Stochastic Variational Inference. \emph{Journal of Machine Learning Research}, 2013\cr
}
\author{
Zhehui Chen, Shiyang Li, Xingguo Li, Tuo Zhao
}
