% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{svi_map}
\alias{svi_map}
\title{Plot US SVI Map}
\usage{
svi_map(data, index)
}
\arguments{
\item{data}{data. It is assumed to have a \code{fips} column.}

\item{index}{string of column name which represents the SVI estimates.}
}
\value{
a US SVI Map.
}
\description{
This function plots US SVI map for selected estimates.
}
\examples{
library(dplyr)
library(purrr)
data("vulnerability")
vulnerability |>
  mnnn_to_na(names(which(map_lgl(vulnerability, is.double))), -999) |>
  rename(fips = FIPS) |>
  svi_map("EPL_POV")
}
