% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_r.R
\docType{data}
\name{tx}
\alias{tx}
\title{A Compositional Data Analysis of Texas County SVI Indices.}
\format{
a data frame with 254 observations
\describe{
\item{county}{The county.}
\item{SPL_THEMES}{The total themes.}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
tx
}
\description{
A Compositional Data Analysis of Texas County SVI Indices.
}
\keyword{datasets}
