% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vulnerability}
\alias{vulnerability}
\title{Social Vulnerability Index per County}
\description{
Contains social vulnerability measures in 4 themes 15 factors: \cr
\enumerate{
\item Socioeconomic: below poverty, unemployed, income, no high school diploma \cr
\item Household Composition/Disability: aged 65 or older, aged 17 or younger, older than age 5 with a disability, single-parent households \cr
\item Minority Status/Language: minority, speaks English "less than well" \cr
\item Housing Type/Transportation: multi-unit structures, mobile homes, crowding, no vehicle, group quarters
}
}
\details{
See
\href{https://svi.cdc.gov/Documents/Data/2018_SVI_Data/SVI2018Documentation.pdf}{CDC SVI 2018 Documentation}
for more information
\describe{
\item{E_}{estimates}
\item{EP_}{percentage of estimates}
\item{EPL_}{percentile percentage of estimates}
\item{M_}{margins of error for estimates}
\item{SPL_}{sum of percentile percentage in the theme}
\item{RPL_}{percentile ranking of the theme}
}
}
\keyword{data}
