% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{cr_interpolate}
\alias{cr_interpolate}
\title{Get the Spectral Signature of Accelerometry Data}
\usage{
cr_interpolate(data, reliable = TRUE)
}
\arguments{
\item{data}{data}

\item{reliable}{default is TRUE}
}
\description{
The spectral signature is calculated by taking the modulus of the
Fourier coefficients of the signal.
}
