% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{mnnn_to_na}
\alias{mnnn_to_na}
\title{Substitute Target Value into NA}
\usage{
mnnn_to_na(data, column, target)
}
\arguments{
\item{data}{a dataframe which needs to change some items into NA.}

\item{column}{column needs to be substituted.}

\item{target}{target value.}
}
\value{
a data frame with target value in selected columns replaced with NA.
}
\description{
This function changes target value in a given data columns to NA.
}
\examples{
data("diabetes")
diabetes |> mnnn_to_na("Crude.Rate", "Unreliable")
}
