% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{mnnn_to_na}
\alias{mnnn_to_na}
\title{Get the Spectral Signature of Accelerometry Data}
\usage{
mnnn_to_na(data, column, target)
}
\arguments{
\item{data}{data}

\item{column}{column needs to be substituted}

\item{target}{target}
}
\description{
The spectral signature is calculated by taking the modulus of the
Fourier coefficients of the signal.
}
