% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{mortality_by_state}
\alias{mortality_by_state}
\title{Get the Spectral Signature of Accelerometry Data}
\usage{
mortality_by_state(data)
}
\arguments{
\item{data}{data}
}
\description{
The spectral signature is calculated by taking the modulus of the
Fourier coefficients of the signal.
}
