% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mortality_map}
\alias{mortality_map}
\title{Plot US Mortality Map}
\usage{
mortality_map(data, colname, level = "County", disease = "Diabetes")
}
\arguments{
\item{data}{data. It is assumed to have a \code{fips} column.}

\item{colname}{string of column name which represents the mortality in the data.}

\item{level}{string, "County" or "State". (Default is \code{County})}

\item{disease}{string of the disease name showing. (Default is \code{Diabetes})}
}
\value{
a US mortality map.
}
\description{
This function plots US mortality map for selected disease, either in county level or state level (should be indicated
in parameter).
}
\examples{
library(dplyr)
data("diabetes")
diabetes |>
  cr_interpolate(reliable = FALSE) |>
  rename(fips = County.Code) |>
  mortality_map("Crude.Rate", "County", "Diabetes")
}
