% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mortality_vs_svi_scatter}
\alias{mortality_vs_svi_scatter}
\title{Scatter Plot of Mortality vs. SVI Estimates}
\usage{
mortality_vs_svi_scatter(data, index, disease = "Diabetes")
}
\arguments{
\item{data}{prepared data. It is assumed to have a \code{MORTALITY} column.}

\item{index}{selected SVI estimates.}

\item{disease}{string of disease name. (Default is \code{Diabetes})}
}
\value{
a scatter plot.
}
\description{
This function shows the scatter plot of mortality versus the selected SVI estimates.
}
\examples{
data("vulnerability")
data("diabetes")
df <- prepare(vulnerability, diabetes, reliable = FALSE)
mortality_vs_svi_scatter(df, "RPL_THEME1", "Diabetes")
}
