% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{prepare}
\alias{prepare}
\title{Prepare Data for Plot and Regression}
\usage{
prepare(vulnerability, mortality, reliable = TRUE)
}
\arguments{
\item{vulnerability}{original SVI data from CDC.}

\item{mortality}{original mortality data from CDC.}

\item{reliable}{if taking into account deaths <= 20. (Default is \code{TRUE})}
}
\value{
a data frame joining SVI data and mortality data.
}
\description{
This function joins SVI data and mortality data from CDC.
}
\examples{
data("vulnerability")
data("diabetes")
prepare(vulnerability, diabetes)
}
