% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{svi_map}
\alias{svi_map}
\title{Get the Spectral Signature of Accelerometry Data}
\usage{
svi_map(data, index)
}
\arguments{
\item{data}{data}

\item{index}{string}
}
\value{
a plot
}
\description{
The spectral signature is calculated by taking the modulus of the
Fourier coefficients of the signal.
}
